Bump minimum required CMake to 3.18
CMake 3.18 introduces the new variable CMAKE_CUDA_ARCHITECTURES and
target property CUDA_ARCHITECTURE:

https://cmake.org/cmake/help/v3.18/variable/CMAKE_CUDA_ARCHITECTURES.html
https://cmake.org/cmake/help/v3.18/prop_tgt/CUDA_ARCHITECTURES.html

These list the required architectures to generate device code for, and
set the required CUDA compiler flags. See e.g.

https://docs.nvidia.com/cuda/cuda-compiler-driver-nvcc/index.html#gpu-generations

To better support selection of CUDA architecture flags, require
CMake 3.18 as the minimum. CMAKE_CUDA_ARCHITECTURES is not set so the
defaults for the compiler will be used until the range of supported
devices is defined.