RANLUX++ for host and device
Copied from ROOT, with support for only one luxury level (p=2048).
The main interface for now is RanluxppDouble, but we can implement
more classes that use less random bits to produce a float.
The added test shows how the state can be transferred between host
and device via a cudaMemcpy and checks that the generated numbers
on host and device are identical.