Optimize RNG workflow with MSC
Multiple Scattering uses quite some random numbers, even if there
is no discrete interaction with final state sampling happening. This
causes problems for the GPU performance because the RNG might need
to advance the state when threads have already diverged. Moving the
call to branch a new RNG state earlier and providing a fresh round
of numbers shows a clear performance improvement of around 20% for
TestEm3 and around 10% for Example13 with cms2018.